Fix : routine 이미지 업로드 안되는 현상 해결
