+++ WEB(BE)/src/controllers/signup.ctrl.js
@@ -56,7 +56,8 @@ const user = {
 			return res.render('alert', {error: '이미 가입된 이메일입니다'});
 		}
 		
-		 res.redirect('/signup/more/' + userId);
+		
+		 // res.redirect('/signup/more/' + userId);
 	},
 	
 	addInfo : async(req, res) => {
